test: profile_hobbies_updater_spec を profile_spec に統合し、孤立 spec を削除
- Profile#update_hobbies_from のテストを profile_spec.rb に移植・削除
  （ファイル名と describe 対象が不一致だったため）
- spec/views/my/share_links/ を削除
  （対応するコントローラー・ビューが存在しない空の pending spec）
