--- spec/views/my/share_links/update.html.erb_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "share_links/update.html.erb", type: :view do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
